% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altcoin_marktanalyse_cmc.R
\name{altcoin_marktanalyse_cmc}
\alias{altcoin_marktanalyse_cmc}
\title{Show details about last week}
\usage{
altcoin_marktanalyse_cmc(circulating = F)
}
\value{
A list with means, the winners, the losers, details about the top 10
and the differences regarding market cap.
}
\description{
Show development for cryptocurrencies with focus on the top 10.
}
\examples{
altcoin_marktanalyse_cmc()
}
\author{
Philipp Giese
}
