% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altcoin_marktanalyse_cmc_daily.R
\name{altcoin_marktanalyse_cmc_daily}
\alias{altcoin_marktanalyse_cmc_daily}
\title{Generate Coin Report about}
\usage{
altcoin_marktanalyse_cmc_daily(number, file_path)
}
\value{
A list with means, the winners, the losers, details about the top 10
and the differences regarding market cap.
}
\description{
Show development for cryptocurrencies with focus on the top 10.
}
\examples{
altcoin_marktanalyse_cmc_daily(2,path)
}
\author{
Philipp Giese
}
