% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gains.R
\name{calculate_gains}
\alias{calculate_gains}
\title{Get gains from cryptocompare}
\usage{
calculate_gains(data)
}
\arguments{
\item{data}{A xts with ohlc-values}
}
\value{
A time series with gains
}
\description{
Calculate the gains of a cryptocurrency with respect to the
closing price within a xts of ohlc-values. Referred_to columns
must be named "close" and "open".
}
\examples{
calculate_gains(calculate_price("CCCAGG","BTC","USD",7))
}
\author{
Philipp Giese
}
