% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multiple_cc.R
\name{calculate_multiple_cc}
\alias{calculate_multiple_cc}
\title{Get prices or gains for several Cryptocurrencies
from cryptocompare}
\usage{
calculate_multiple_cc(exchange = "CCCAGG", currencies_from,
  currency_to = "USD", days, gains)
}
\arguments{
\item{exchange}{A string defining the exchange}

\item{currencies_from}{A vector of strings defining the
analyzed currencies}

\item{currency_to}{A string defining the base currency}

\item{days}{An integer defining past x days analyzed}

\item{gains}{boolean, T for gains, F for OHLC-values}
}
\value{
A time series with gains (when gains = T)
or ohlc-values and volume (when gains = F)
}
\description{
Read the OHLC-Values + Volume or the gains for a list of crypto-
currencies.
}
\examples{
calculate_multiple_cc("CCCAGG",c("BTC","LTC","test","CRAIG"),"USD",7,T)
}
\author{
Philipp Giese
}
