% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_price.R
\name{calculate_price}
\alias{calculate_price}
\title{Get price from cryptocompare}
\usage{
calculate_price(exchange, currency_from, currency_to, days)
}
\arguments{
\item{exchange}{A string defining the exchange}

\item{currency_from}{A string defining the analyzed currency}

\item{currency_to}{A string defining the base currency}

\item{days}{An integer defining past x days analyzed}
}
\value{
A time series with ohlc-values and volume
}
\description{
Read the OHLC-values for a distinct cryptocurrency_from in units
of cryptocurrency_to at the exchange for the last days
}
\examples{
calculate_price("CCCAGG","BTC","USD",7)
}
\author{
Philipp Giese
}
