% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_switch_cc.R
\name{calculate_switch_cc}
\alias{calculate_switch_cc}
\title{Get price or gains from cryptocompare}
\usage{
calculate_switch_cc(exchange, currency_from, currency_to, days, gains)
}
\arguments{
\item{exchange}{A string defining the exchange}

\item{currency_from}{A string defining the analyzed currency}

\item{currency_to}{A string defining the base currency}

\item{days}{An integer defining past x days analyzed}

\item{gains}{boolean, T for gains, F for OHLC-values}
}
\value{
A time series with ohlc-values and volume
}
\description{
Read the OHLC-values or gains for a distinct cryptocurrency_from
in units of cryptocurrency_to at the exchange for the last days
}
\examples{
calculate_switch_cc("CCCAGG","BTC","USD",7,T)
}
\author{
Philipp Giese
}
