% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_data.R
\name{cleanup_data}
\alias{cleanup_data}
\title{Clean Data from outliers}
\usage{
cleanup_data(returns_clean, k_value)
}
\arguments{
\item{returns_clean}{A xts with several columns consisting of data}

\item{k_value}{A number defining which outliers should be neglected}
}
\value{
A xts with less outliers
}
\description{
Clean a time series from outliers. The definition of an outlier
is a value who is outside an interval mean(data)+/-k_value*sd(data)
with data being a column of a time series and k_value an integer.
}
\examples{
data<-calculate_mutliple_cc("CCCAGG",c("BTC","LTC","DASH"),"USD",7,T)
cleanup_data(data,5)
}
\author{
Philipp Giese
}
