% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cet_coins_cc.R
\name{get_coins_cc}
\alias{get_coins_cc}
\title{Get cryptocurrencies from cryptocompare}
\usage{
get_coins_cc()
}
\value{
List of coins on cryptocompare
}
\description{
Get a vector with the cryptocurrency-IDs
from cryptocompare
}
\examples{
get_coins_cc()
}
\author{
Philipp Giese
}
