% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_news_ofx.R
\name{get_news_ofx}
\alias{get_news_ofx}
\title{Get news from onchainfx}
\usage{
get_news_ofx(coin = NULL)
}
\arguments{
\item{coin}{optional specification for coin (default empty)}
}
\value{
A list with fifty news regarding the specified cryptocurr
}
\description{
Show a curated list of the top last 50 news by the OnchainFx-Team
}
\examples{
get_news_ofx()
}
\author{
Philipp Giese
}
