% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portf_dev.R
\name{get_portf_dev}
\alias{get_portf_dev}
\title{Show Portfolio development}
\usage{
get_portf_dev(returns, portfolio, date_margins)
}
\arguments{
\item{returns}{A xts with several columns consisting of data}

\item{portfolio}{A vector defining the portfolio weights}

\item{date_margins}{A string specifying the start- and enddates}
}
\value{
A xts with development of portfolio
}
\description{
Show the development of a weighted portfolio
}
\examples{
data<-calculate_mutliple_cc("CCCAGG",c("BTC","LTC","DASH"),"USD",7,T)
portf<-c(0.3,0.4,0.3)
get_portf_dev(data,portf,"201810/")
}
\author{
Philipp Giese
}
