% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sel_data_cg.R
\name{get_sel_data_cg}
\alias{get_sel_data_cg}
\title{Get all available historical data for a coin}
\usage{
get_sel_data_cg(coin, coin_basis, days, sel)
}
\arguments{
\item{coin}{coin name eg "bitcoin"}

\item{coin_basis}{base currency eg "usd"}

\item{days}{number of days, eg number or "max"}

\item{sel}{selection of Data: 1 for price, 2 for market cap and 3 for volume.}
}
\value{
A xts either of price, market cap and volume
}
\description{
Get daily price, market cap and volume for a specific crypto-
currency from coingecko
}
\examples{
get_sel_data_cg("bitcoin","usd","max",1)
}
\author{
Philipp Giese
}
