% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sup_resi.R
\name{get_sup_resi}
\alias{get_sup_resi}
\title{Get Supports and Resistances}
\usage{
get_sup_resi(ohlc, max_past, width)
}
\arguments{
\item{ohlc}{an xts with the close, high, low, open and volume-Vals}

\item{max_past}{An integer defining past x days analyzed}

\item{width}{binning width}
}
\value{
a density plot with the maxima and a vector with all maxima
}
\description{
Generates a density plot wit a chosen width of the close-values of a ohlc-xts and lists the maxima
}
\examples{
get_sup_resi(btcecho:::calculate_price("CCCAGG","BNB","USD",2000),"20180301/",0.1)
}
\author{
Philipp Giese
}
