% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_portf.R
\name{pie_portf}
\alias{pie_portf}
\title{Pie Plot of Portfolio}
\usage{
pie_portf(portfolio, returns)
}
\arguments{
\item{portfolio}{A vector defining the portfolio weights}

\item{returns}{A xts with several columns consisting of data}
}
\value{
A xts with development of portfolio
}
\description{
Make a pieplot of the current portfolio
}
\examples{
data<-calculate_mutliple_cc("CCCAGG",c("BTC","LTC","DASH"),"USD",7,T)
portf<-c(0.3,0.4,0.3)
get_portf_dev(data,portf,"201810/")
}
\author{
Philipp Giese
}
