% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ohlc_complete.R
\name{plot_ohlc_complete}
\alias{plot_ohlc_complete}
\title{Plot OHLC complete}
\usage{
plot_ohlc_complete(ohlc, ymin, ymax, max_past)
}
\arguments{
\item{ohlc}{an xts with the close, high, low, open and volume-Vals}

\item{ymin}{minimal value on y-axis}

\item{ymax}{maximal value on y-axis}

\item{max_past}{An integer defining past x days analyzed}
}
\value{
A nifty plot with MA140, MA980 and MA1400 and other indicators
}
\description{
Generate a nice OHLC/candlestick plot with different moving averages
plus a MACD, a RSI and an Aroon-indicator
To do: yaxis ticks and format of indicators
}
\examples{
plot_ohlc_complete(btcecho:::calculate_price("CCCAGG","BNB","USD",2000),0,20,365)
}
\author{
Philipp Giese
}
