% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rendite.R
\name{plot_rendite}
\alias{plot_rendite}
\title{Plot Profits}
\usage{
plot_rendite(returns, portfolio, returns_safe, date_margin, label,
  legend_pos, lwd_thick, lwd_thin)
}
\arguments{
\item{returns}{A xts with several columns consisting of data}

\item{portfolio}{A vector defining the portfolio weights}

\item{date_margin}{A string specifying the start- and enddates}

\item{label}{How to define the examined portfolio within the legend}

\item{legend_pos}{where to set the legend}

\item{lwd_tick}{thickness of main line}

\item{lwd_tin}{thickness of other lines}
}
\value{
A xts with development of portfolio
}
\description{
Plot the portfolio-development in comparison to BTC, ETH and
XRP within the examined timeframe
}
\examples{
data<-calculate_mutliple_cc("CCCAGG",c("BTC","LTC","DASH"),"USD",7,T)
portf<-c(0.3,0.4,0.3)
get_portf_dev(data,portf,"201810/")
}
\author{
Philipp Giese
}
