% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rendite_artikel.R
\name{plot_rendite_artikel}
\alias{plot_rendite_artikel}
\title{Plot Returns of different portfolios}
\usage{
plot_rendite_artikel(date_start, returns_safe, portf, rendite, label,
  thick1 = 2, thick2 = 4)
}
\arguments{
\item{date_start}{A xts with several columns consisting of data}

\item{returns_safe}{A vector defining the portfolio weights}

\item{portf}{A xts with several columns consisting of data}

\item{rendite}{A vector defining the portfolio weights}

\item{label}{A xts with several columns consisting of data}

\item{thick1}{thickness of normal lines}

\item{thick2}{thickness of portfolio}
}
\value{
A plot with development of tree portfolios
}
\description{
Plot one portfolio-development in comparison to two other ones
}
\examples{
returns_safe<-calculate_mutliple_cc("CCCAGG",c("BTC","LTC","DASH"),"USD",7,T)
portf_safe<-c(0.3,0.4,0.3)
}
\author{
Philipp Giese
}
