% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rendite_comp.R
\name{plot_rendite_comp}
\alias{plot_rendite_comp}
\title{Plot Returns of different portfolios}
\usage{
plot_rendite_comp(returns_safe, portf_safe, returns_medium, portf_medium,
  returns_risky, portf_risky, date_margin, legend_pos)
}
\arguments{
\item{returns_safe}{A xts with several columns consisting of data}

\item{portf_safe}{A vector defining the portfolio weights}

\item{returns_medium}{A xts with several columns consisting of data}

\item{portf_medium}{A vector defining the portfolio weights}

\item{returns_risky}{A xts with several columns consisting of data}

\item{portf_risky}{A vector defining the portfolio weights}

\item{date_margin}{A string specifying the start- and enddates}

\item{legend_pos}{where to set the legend}
}
\value{
A plot with development of tree portfolios
}
\description{
Plot one portfolio-development in comparison to two other ones
}
\examples{
returns_safe<-calculate_mutliple_cc("CCCAGG",c("BTC","LTC","DASH"),"USD",7,T)
portf_safe<-c(0.3,0.4,0.3)
returns_medium<-calculate_mutliple_cc("CCCAGG",c("ADA","EOS","ZRX"),"USD",7,T)
portf_medium<-c(0.2,0.4,0.4)
returns_risky<-calculate_mutliple_cc("CCCAGG",c("PAY","BNB","ZCL"),"USD",7,T)
portf_risky<-c(0.2,0.4,0.4)
plot_rendite_comp(returns_safe,portf_safe,returns_medium,portf_medium,returns_risky,portf_risky,"20180915/","bottomleft")
}
\author{
Philipp Giese
}
