% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotxts.R
\name{plotxts}
\alias{plotxts}
\title{Plot xts without date range}
\usage{
plotxts(x, y = NULL, ..., subset = "", panels = NULL,
  multi.panel = FALSE, col = 1:8, up.col = NULL, dn.col = NULL,
  bg = "#FFFFFF", type = "l", lty = 1, lwd = 2, lend = 1,
  main = deparse(substitute(x)), observation.based = FALSE,
  ylim = NULL, yaxis.same = TRUE, yaxis.left = TRUE,
  yaxis.right = TRUE, major.ticks = "months", minor.ticks = NULL,
  grid.ticks.on = "months", grid.ticks.lwd = 1, grid.ticks.lty = 1,
  grid.col = "darkgray", labels.col = "#333333",
  format.labels = TRUE, grid2 = "#F5F5F5", legend.loc = NULL)
}
\arguments{
\item{...}{all arguments as known with plot.xts}
}
\value{
The same as plot.xts without the date range
}
\description{
a small tweak of plot.xts as presented by Joshua Ulrich
on https://stackoverflow.com/questions/50051076/changes-in-plotting-an-xts-object/50051183#50051183
}
\author{
Ross Bennett (Tweak by Joshua Ulrich)
}
