% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_update.R
\name{port_update}
\alias{port_update}
\title{Plot Returns of different portfolios}
\usage{
port_update(date_margin, returns, offset, portfolio)
}
\arguments{
\item{date_margin}{A string specifying the start- and enddates}

\item{returns}{A xts with several columns consisting of data}

\item{offset}{A vector defining the portfolio weights}

\item{portfolio}{A vector defining the portfolio weights}
}
\value{
A xts with development of portfolio
}
\description{
Plot one portfolio-development in comparison to two other ones
}
\examples{
returns_safe<-calculate_mutliple_cc("CCCAGG",c("BTC","LTC","DASH"),"USD",7,T)
portf_safe<-c(0.3,0.4,0.3)
returns_medium<-calculate_mutliple_cc("CCCAGG",c("ADA","EOS","ZRX"),"USD",7,T)
portf_medium<-c(0.2,0.4,0.4)
returns_risky<-calculate_mutliple_cc("CCCAGG",c("PAY","BNB","ZCL"),"USD",7,T)
portf_risky<-c(0.2,0.4,0.4)
plot_rendite_comp(returns_safe,portf_safe,returns_medium,portf_medium,returns_risky,portf_risky,"20180915/","bottomleft")
}
\author{
Philipp Giese
}
