% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_portf_excel.R
\name{read_portf_excel}
\alias{read_portf_excel}
\title{Read portfolio development from Excel file}
\usage{
read_portf_excel(file, sheet)
}
\arguments{
\item{file}{the path to the xlsx file}

\item{sheet}{the sheet where the data is to be found}
}
\value{
A time series with portfolio development
}
\description{
Read the portfolio development from an excel file. The excel sheet
should have the following structure:
<empty> <Asset_1> <Asset_2>...
<date> <weight_Asset_1> <weight_Asset_2>...
...
}
\examples{
calculate_gains(calculate_price("CCCAGG","BTC","USD",7))

}
\author{
Philipp Giese
}
