% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermocap.R
\name{thermocap}
\alias{thermocap}
\title{Get thermocap for bitcoin}
\usage{
thermocap()
}
\value{
A time series with portfolio development
}
\description{
Thermocap is an alternative metric for evaluating cryptocurrencies.
Unlike market cap thermocap is a sum over the prices when the coins
are generated.
}
\examples{
thermocap()
}
\author{
Philipp Giese
}
